requirements.txt
functions-framework==3.*
google-cloud-resource-manager~=1.10
google-cloud-billing~=1.11
google-cloud-billing-budgets~=1.11
logwrap~=11.0